Align data_dir() logic
This change should have no impact to the current CLI logic, because
dirs::data_dir() is the same as dir::data_local_dir(), and we're not
using platform::data_dir() on Windows yet. But this change aligns the
implementation of platform::data_dir() with other methods, making it
seamless when we have WSL-native instances on Windows in the future.